--- src/components/organisms/modal/Modal.tsx
@@ -1,14 +1,0 @@
-import ModalView from "../../molecules/modal/ModalView";
-import { modalProps } from "@/types/Modal";
-import React from "react";
-
-const Modal = ({ info, yesBtn, noBtn }: modalProps) => {
-  return (
-    <div>d</div>
-    // <ModalView>
-    //   <DefaultModal info={info} yesBtn={yesBtn} noBtn={noBtn} />
-    // </ModalView>
-  );
-};
-
-export default Modal;
